New structure

+++ src/services/imageKit.ts
@@ -7,12 +7,11 @@ function uploadImage(base64: string, fileName: string, cdnFolder: string) {
     useUniqueFileName: false,
     folder: cdnFolder
   }, function (error, result) {
-    if (error) {
+    if (!result) {
       console.log(error)
-      return 'error';
-    } else {
-      return result.url;
-    };
+      return error?.message;
+    }
+    return result.url;
   });
 };
 
@@ -21,7 +20,10 @@ async function updateImages(fileToDelete: string, base64: string, fileName: stri
     type: 'file',
     searchQuery: `name = ${fileToDelete}`
   }, async (error, result) => {
-    if (error) console.log(error);
+    if (!result) {
+      console.log(error);
+      return error?.message;
+    }
 
     result[0] ? await imageKit.deleteFile(result[0].fileId).then(async res => {
       await imageKit.purgeCache(result[0].url)
@@ -34,9 +36,9 @@ async function updateImages(fileToDelete: string, base64: string, fileName: stri
     useUniqueFileName: false,
     folder: cdnFolder
   }, function (error, result) {
-    if (error) {
+    if (!result) {
       console.log(error)
-      return 'error';
+      return error?.message;
     } else {
       return result.url;
     };
@@ -49,7 +51,10 @@ async function deleteFile(fileToDelete: string, path: string) {
     searchQuery: `name = ${fileToDelete}`,
     path: path
   }, async (error, result) => {
-    if (error) console.log(error);
+    if (!result) {
+      console.log(error?.message);
+      return error?.message;
+    };
 
     result[0] ? await imageKit.deleteFile(result[0].fileId).then(async res => {
       await imageKit.purgeCache(result[0].url)
